backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.7.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
bcrypt==3.2.0
email-validator
requests
uvloop; sys_platform != "win32"
httptools; sys_platform != "win32"
